refactor(home): move health-score helpers + interface to shared/lib; drop redundant score prop
Fixes the shared/ → features/ import direction by relocating HealthStats,
calculateHealthStats, and calculateHealthScore from
features/ai-intelligence/components/fleet-health-summary.tsx into a new
shared/lib/health-score.ts. fleet-health-summary.tsx re-exports them so
existing import paths continue to work.

Drops the redundant `score` prop from HealthScoreCard — the component now
derives it internally from `stats` via calculateHealthScore, removing the
risk of callers passing a score that disagrees with the stats. Updates
home.tsx and fleet-health-summary.tsx to the new single-prop API.
